requirements.txt
requests>=2.28
cryptography>=41.0
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0